requirements.txt
pyTelegramBotAPI>=4.0.0
shazamio>=0.0.5
aiohttp>=3.8.0
asyncio
requests>=2.25.0
python-dotenv>=0.19.0

# For YouTube downloading
pytube>=15.0.0

# For Instagram downloading
instaloader>=4.9.0

# For metadata editing
mutagen>=1.46.0